+++ CHANGELOG.md
@@ -6,6 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.32] - 2020-07-08
+### 🐛 Bug fixes
+- Fixes to active-referral ([#238](https://github.com/whirli/whirli-client/pull/239))
+
+## [0.0.31] - 2020-07-08
+### 🐛 Bug fixes
+- Fix get method and referrals interface ([#238](https://github.com/whirli/whirli-client/pull/238))
+
+## [0.0.30] - 2020-07-08
+### 🚀 Features
+- Add active-referral endpoint ([#235](https://github.com/whirli/whirli-client/pull/235))
+
 ## [0.0.29] - 2020-07-06
 ### 🐛 Bug fixes
 - Change slug name from hyphenated to not on public product-lists to fix endpoint not registering it ([#232](https://github.com/whirli/whirli-client/pull/232))